+++ cmd/zetae2e/local/local.go
@@ -318,7 +318,11 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 			logger.Print("❌ solana client is nil, maybe solana rpc is not set")
 			os.Exit(1)
 		}
-		eg.Go(solanaTestRoutine(conf, deployerRunner, verbose, e2etests.TestSolanaDepositName))
+		solanaTests := []string{
+			e2etests.TestSolanaDepositName,
+			e2etests.TestSolanaWithdrawName,
+		}
+		eg.Go(solanaTestRoutine(conf, deployerRunner, verbose, solanaTests...))
 	}
 
 	// while tests are executed, monitor blocks in parallel to check if system txs are on top and they have biggest priority