+++ zetaclient/chains/bitcoin/observer/outbound.go
@@ -23,12 +23,6 @@ import (
 	"github.com/zeta-chain/zetacore/zetaclient/zetacore"
 )
 
-// GetTxID returns a unique id for outbound tx
-func (ob *Observer) GetTxID(nonce uint64) string {
-	tssAddr := ob.TSS().BTCAddress()
-	return fmt.Sprintf("%d-%s-%d", ob.Chain().ChainId, tssAddr, nonce)
-}
-
 // WatchOutbound watches Bitcoin chain for outgoing txs status
 // TODO(revamp): move ticker functions to a specific file
 // TODO(revamp): move into a separate package
@@ -66,7 +60,7 @@ func (ob *Observer) WatchOutbound(ctx context.Context) error {
 			}
 			for _, tracker := range trackers {
 				// get original cctx parameters
-				outboundID := ob.GetTxID(tracker.Nonce)
+				outboundID := ob.OutboundID(tracker.Nonce)
 				cctx, err := ob.ZetacoreClient().GetCctxByNonce(ctx, chainID, tracker.Nonce)
 				if err != nil {
 					ob.logger.Outbound.Info().
@@ -120,13 +114,11 @@ func (ob *Observer) WatchOutbound(ctx context.Context) error {
 	}
 }
 
-// IsOutboundProcessed returns isIncluded(or inMempool), isConfirmed, Error
-// TODO(revamp): rename as it vote the outbound and doesn't only check if outbound is processed
-func (ob *Observer) IsOutboundProcessed(
+// VoteOutboundIfConfirmed checks outbound status and returns (continueKeysign, error)
+func (ob *Observer) VoteOutboundIfConfirmed(
 	ctx context.Context,
 	cctx *crosschaintypes.CrossChainTx,
-	logger zerolog.Logger,
-) (bool, bool, error) {
+) (bool, error) {
 	const (
 		// not used with Bitcoin
 		outboundGasUsed  = 0
@@ -141,8 +133,8 @@ func (ob *Observer) IsOutboundProcessed(
 	nonce := cctx.GetCurrentOutboundParam().TssNonce
 
 	// get broadcasted outbound and tx result
-	outboundID := ob.GetTxID(nonce)
-	logger.Info().Msgf("IsOutboundProcessed %s", outboundID)
+	outboundID := ob.OutboundID(nonce)
+	ob.Logger().Outbound.Info().Msgf("VoteOutboundIfConfirmed %s", outboundID)
 
 	ob.Mu().Lock()
 	txnHash, broadcasted := ob.broadcastedTx[outboundID]
@@ -151,7 +143,7 @@ func (ob *Observer) IsOutboundProcessed(
 
 	if !included {
 		if !broadcasted {
-			return false, false, nil
+			return true, nil
 		}
 		// If the broadcasted outbound is nonce 0, just wait for inclusion and don't schedule more keysign
 		// Schedule more than one keysign for nonce 0 can lead to duplicate payments.
@@ -159,16 +151,16 @@ func (ob *Observer) IsOutboundProcessed(
 		// prevents double spending of same UTXO. However, for nonce 0, we don't have a prior nonce (e.g., -1)
 		// for the signer to check against when making the payment. Signer treats nonce 0 as a special case in downstream code.
 		if nonce == 0 {
-			return true, false, nil
+			return false, nil
 		}
 
 		// Try including this outbound broadcasted by myself
 		txResult, inMempool := ob.checkIncludedTx(ctx, cctx, txnHash)
 		if txResult == nil { // check failed, try again next time
-			return false, false, nil
+			return true, nil
 		} else if inMempool { // still in mempool (should avoid unnecessary Tss keysign)
-			ob.logger.Outbound.Info().Msgf("IsOutboundProcessed: outbound %s is still in mempool", outboundID)
-			return true, false, nil
+			ob.logger.Outbound.Info().Msgf("VoteOutboundIfConfirmed: outbound %s is still in mempool", outboundID)
+			return false, nil
 		}
 		// included
 		ob.setIncludedTx(nonce, txResult)
@@ -176,9 +168,9 @@ func (ob *Observer) IsOutboundProcessed(
 		// Get tx result again in case it is just included
 		res = ob.getIncludedTx(nonce)
 		if res == nil {
-			return false, false, nil
+			return true, nil
 		}
-		ob.logger.Outbound.Info().Msgf("IsOutboundProcessed: setIncludedTx succeeded for outbound %s", outboundID)
+		ob.logger.Outbound.Info().Msgf("VoteOutboundIfConfirmed: setIncludedTx succeeded for outbound %s", outboundID)
 	}
 
 	// It's safe to use cctx's amount to post confirmation because it has already been verified in observeOutbound()
@@ -187,22 +179,24 @@ func (ob *Observer) IsOutboundProcessed(
 		ob.logger.Outbound.Debug().
 			Int64("currentConfirmations", res.Confirmations).
 			Int64("requiredConfirmations", ob.ConfirmationsThreshold(amountInSat)).
-			Msg("IsOutboundProcessed: outbound not confirmed yet")
+			Msg("VoteOutboundIfConfirmed: outbound not confirmed yet")
 
-		return true, false, nil
+		return false, nil
 	}
 
 	// Get outbound block height
 	blockHeight, err := rpc.GetBlockHeightByHash(ob.btcClient, res.BlockHash)
 	if err != nil {
-		return true, false, errors.Wrapf(
+		return false, errors.Wrapf(
 			err,
-			"IsOutboundProcessed: error getting block height by hash %s",
+			"VoteOutboundIfConfirmed: error getting block height by hash %s",
 			res.BlockHash,
 		)
 	}
 
-	logger.Debug().Msgf("Bitcoin outbound confirmed: txid %s, amount %s\n", res.TxID, amountInSat.String())
+	ob.Logger().
+		Outbound.Debug().
+		Msgf("Bitcoin outbound confirmed: txid %s, amount %s\n", res.TxID, amountInSat.String())
 
 	signer := ob.ZetacoreClient().GetKeys().GetOperatorAddress()
 
@@ -236,12 +230,16 @@ func (ob *Observer) IsOutboundProcessed(
 	}
 
 	if err != nil {
-		logger.Error().Err(err).Fields(logFields).Msg("IsOutboundProcessed: error confirming bitcoin outbound")
+		ob.Logger().
+			Outbound.Error().
+			Err(err).
+			Fields(logFields).
+			Msg("VoteOutboundIfConfirmed: error confirming bitcoin outbound")
 	} else if zetaHash != "" {
-		logger.Info().Fields(logFields).Msgf("IsOutboundProcessed: confirmed Bitcoin outbound")
+		ob.Logger().Outbound.Info().Fields(logFields).Msgf("VoteOutboundIfConfirmed: confirmed Bitcoin outbound")
 	}
 
-	return true, true, nil
+	return false, nil
 }
 
 // SelectUTXOs selects a sublist of utxos to be used as inputs.
@@ -438,7 +436,7 @@ func (ob *Observer) checkIncludedTx(
 	cctx *crosschaintypes.CrossChainTx,
 	txHash string,
 ) (*btcjson.GetTransactionResult, bool) {
-	outboundID := ob.GetTxID(cctx.GetCurrentOutboundParam().TssNonce)
+	outboundID := ob.OutboundID(cctx.GetCurrentOutboundParam().TssNonce)
 	hash, getTxResult, err := rpc.GetTxResultByHash(ob.btcClient, txHash)
 	if err != nil {
 		ob.logger.Outbound.Error().Err(err).Msgf("checkIncludedTx: error GetTxResultByHash: %s", txHash)
@@ -467,7 +465,7 @@ func (ob *Observer) checkIncludedTx(
 // setIncludedTx saves included tx result in memory
 func (ob *Observer) setIncludedTx(nonce uint64, getTxResult *btcjson.GetTransactionResult) {
 	txHash := getTxResult.TxID
-	outboundID := ob.GetTxID(nonce)
+	outboundID := ob.OutboundID(nonce)
 
 	ob.Mu().Lock()
 	defer ob.Mu().Unlock()
@@ -496,17 +494,17 @@ func (ob *Observer) setIncludedTx(nonce uint64, getTxResult *btcjson.GetTransact
 func (ob *Observer) getIncludedTx(nonce uint64) *btcjson.GetTransactionResult {
 	ob.Mu().Lock()
 	defer ob.Mu().Unlock()
-	return ob.includedTxResults[ob.GetTxID(nonce)]
+	return ob.includedTxResults[ob.OutboundID(nonce)]
 }
 
 // removeIncludedTx removes included tx from memory
 func (ob *Observer) removeIncludedTx(nonce uint64) {
 	ob.Mu().Lock()
 	defer ob.Mu().Unlock()
-	txResult, found := ob.includedTxResults[ob.GetTxID(nonce)]
+	txResult, found := ob.includedTxResults[ob.OutboundID(nonce)]
 	if found {
 		delete(ob.includedTxHashes, txResult.TxID)
-		delete(ob.includedTxResults, ob.GetTxID(nonce))
+		delete(ob.includedTxResults, ob.OutboundID(nonce))
 	}
 }