+++ zetaclient/chains/evm/signer/signer.go
@@ -26,7 +26,6 @@ import (
 	"github.com/zeta-chain/zetacore/pkg/constant"
 	crosschainkeeper "github.com/zeta-chain/zetacore/x/crosschain/keeper"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
-	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 	"github.com/zeta-chain/zetacore/zetaclient/chains/base"
 	"github.com/zeta-chain/zetacore/zetaclient/chains/evm"
 	"github.com/zeta-chain/zetacore/zetaclient/chains/evm/observer"
@@ -48,7 +47,7 @@ const (
 )
 
 var (
-	_ interfaces.ChainSigner = &Signer{}
+	_ interfaces.ChainSigner = (*Signer)(nil)
 
 	// zeroValue is for outbounds that carry no ETH (gas token) value
 	zeroValue = big.NewInt(0)
@@ -75,9 +74,6 @@ type Signer struct {
 
 	// er20CustodyAddress is the address of the ERC20Custody contract
 	er20CustodyAddress ethcommon.Address
-
-	// outboundHashBeingReported is a map of outboundHash being reported
-	outboundHashBeingReported map[string]bool
 }
 
 // NewSigner creates a new EVM signer
@@ -114,14 +110,13 @@ func NewSigner(
 	}
 
 	return &Signer{
-		Signer:                    baseSigner,
-		client:                    client,
-		ethSigner:                 ethSigner,
-		zetaConnectorABI:          connectorABI,
-		erc20CustodyABI:           custodyABI,
-		zetaConnectorAddress:      zetaConnectorAddress,
-		er20CustodyAddress:        erc20CustodyAddress,
-		outboundHashBeingReported: make(map[string]bool),
+		Signer:               baseSigner,
+		client:               client,
+		ethSigner:            ethSigner,
+		zetaConnectorABI:     connectorABI,
+		erc20CustodyABI:      custodyABI,
+		zetaConnectorAddress: zetaConnectorAddress,
+		er20CustodyAddress:   erc20CustodyAddress,
 	}, nil
 }
 
@@ -139,6 +134,12 @@ func (signer *Signer) SetERC20CustodyAddress(addr ethcommon.Address) {
 	signer.er20CustodyAddress = addr
 }
 
+// SetGatewayAddress sets the gateway address
+func (signer *Signer) SetGatewayAddress(_ string) {
+	// Note: do nothing for now
+	// gateway address will be needed in the future contract architecture
+}
+
 // GetZetaConnectorAddress returns the zeta connector address
 func (signer *Signer) GetZetaConnectorAddress() ethcommon.Address {
 	signer.Lock()
@@ -151,6 +152,13 @@ func (signer *Signer) GetERC20CustodyAddress() ethcommon.Address {
 	signer.Lock()
 	defer signer.Unlock()
 	return signer.er20CustodyAddress
+}
+
+// GetGatewayAddress returns the gateway address
+func (signer *Signer) GetGatewayAddress() string {
+	// Note: return empty string for now
+	// gateway address will be needed in the future contract architecture
+	return ""
 }
 
 // Sign given data, and metadata (gas, nonce, etc)
@@ -354,28 +362,26 @@ func (signer *Signer) TryProcessOutbound(
 	zetacoreClient interfaces.ZetacoreClient,
 	height uint64,
 ) {
-	app, err := zctx.FromContext(ctx)
-	if err != nil {
-		signer.Logger().Std.Error().Err(err).Msg("error getting app context")
-		return
-	}
-
-	logger := signer.Logger().Std.With().
-		Str("outboundID", outboundID).
-		Str("SendHash", cctx.Index).
-		Logger()
-	logger.Info().Msgf("start processing outboundID %s", outboundID)
-	logger.Info().Msgf(
-		"EVM Chain TryProcessOutbound: %s, value %d to %s",
-		cctx.Index,
-		cctx.GetCurrentOutboundParam().Amount.BigInt(),
-		cctx.GetCurrentOutboundParam().Receiver,
-	)
-
+	// end outbound process on panic
 	defer func() {
 		outboundProc.EndTryProcess(outboundID)
+		if err := recover(); err != nil {
+			signer.Logger().Std.Error().Msgf("EVM TryProcessOutbound: %s, caught panic error: %v", cctx.Index, err)
+		}
 	}()
+
+	// prepare logger
+	params := cctx.GetCurrentOutboundParam()
+	logger := signer.Logger().Std.With().
+		Str("method", "TryProcessOutbound").
+		Int64("chain", signer.Chain().ChainId).
+		Uint64("nonce", params.TssNonce).
+		Str("cctx", cctx.Index).
+		Logger()
+
 	myID := zetacoreClient.GetKeys().GetOperatorAddress()
+	logger.Info().
+		Msgf("EVM TryProcessOutbound: %s, value %d to %s", cctx.Index, params.Amount.BigInt(), params.Receiver)
 
 	evmObserver, ok := chainObserver.(*observer.Observer)
 	if !ok {
@@ -448,7 +454,7 @@ func (signer *Signer) TryProcessOutbound(
 			logger.Warn().Err(err).Msg(ErrorMsg(cctx))
 			return
 		}
-	} else if IsSenderZetaChain(cctx, zetacoreClient, &crossChainflags) {
+	} else if IsSenderZetaChain(cctx, zetacoreClient) {
 		switch cctx.InboundParams.CoinType {
 		case coin.CoinType_Gas:
 			logger.Info().Msgf(
@@ -654,15 +660,6 @@ func (signer *Signer) SignERC20WithdrawTx(ctx context.Context, txData *OutboundD
 	return tx, nil
 }
 
-// Exported for unit tests
-
-// GetReportedTxList returns a list of outboundHash being reported
-// TODO: investigate pointer usage
-// https://github.com/zeta-chain/node/issues/2084
-func (signer *Signer) GetReportedTxList() *map[string]bool {
-	return &signer.outboundHashBeingReported
-}
-
 // EvmClient returns the EVM RPC client
 func (signer *Signer) EvmClient() interfaces.EVMRPCClient {
 	return signer.client
@@ -679,10 +676,9 @@ func (signer *Signer) EvmSigner() ethtypes.Signer {
 func IsSenderZetaChain(
 	cctx *types.CrossChainTx,
 	zetacoreClient interfaces.ZetacoreClient,
-	flags *observertypes.CrosschainFlags,
 ) bool {
 	return cctx.InboundParams.SenderChainId == zetacoreClient.Chain().ChainId &&
-		cctx.CctxStatus.Status == types.CctxStatus_PendingOutbound && flags.IsOutboundEnabled
+		cctx.CctxStatus.Status == types.CctxStatus_PendingOutbound
 }
 
 // ErrorMsg returns a error message for SignOutbound failure with cctx data
@@ -759,22 +755,18 @@ func (signer *Signer) reportToOutboundTracker(
 	outboundHash string,
 	logger zerolog.Logger,
 ) {
-	// skip if already being reported
-	signer.Lock()
-	defer signer.Unlock()
-	if _, found := signer.outboundHashBeingReported[outboundHash]; found {
+	// set being reported flag to avoid duplicate reporting
+	alreadySet := signer.Signer.SetBeingReportedFlag(outboundHash)
+	if alreadySet {
 		logger.Info().
 			Msgf("reportToOutboundTracker: outboundHash %s for chain %d nonce %d is being reported", outboundHash, chainID, nonce)
 		return
 	}
-	signer.outboundHashBeingReported[outboundHash] = true // mark as being reported
 
 	// report to outbound tracker with goroutine
 	go func() {
 		defer func() {
-			signer.Lock()
-			delete(signer.outboundHashBeingReported, outboundHash)
-			signer.Unlock()
+			signer.Signer.ClearBeingReportedFlag(outboundHash)
 		}()
 
 		// try monitoring tx inclusion status for 10 minutes